some more unsigned ints, unifying std:string::npos handling, refs #2495
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@21201 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/mesosim/MESegment.h
@@ -133,7 +133,7 @@ public:
 
     /// @brief return the number of queues
     inline int numQueues() const {
-        return myCarQues.size();
+        return (int)myCarQues.size();
     }
     /** @brief Returns the cars in the queue with the given index for visualization
      * @return the Queue (XXX not thread-safe!)
